# server/openapi.yaml
openapi: 3.1.1
info:
  title: Chainsaw Voting API
  version: 1.0.0
  description: API for campaign setup, voting, results, and authentication.
...
paths:
...
  /campaigns/{campaign_id}/roles:
    parameters:
      - name: campaign_id
        in: path
        required: true
        schema:
          type: string
          format: number
          pattern: "^[1-9][0-9]*$"

    post:
      tags:
        - Roles
      summary: Create a campaign role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "201":
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRoleResponse"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "404":
          description: Campaign not found

    get:
      tags:
        - Roles
      summary: Get all roles in a campaign
      responses:
        "200":
          description: Roles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          description: Unauthorized
        "404":
          description: Campaign not found
  /campaigns/{campaign_id}/roles/{role_id}:
    parameters:
      - name: campaign_id
        in: path
        required: true
        schema:
          type: string
          format: number
          pattern: "^[1-9][0-9]*$"

      - name: role_id
        in: path
        required: true
        schema:
          type: string
          format: number
          pattern: "^[1-9][0-9]*$"

    get:
      tags:
        - Roles
      summary: Get a campaign role
      responses:
        "200":
          description: Role retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "401":
          description: Unauthorized
        "404":
          description: Role not found

    patch:
      tags:
        - Roles
      summary: Update a campaign role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleRequest"
      responses:
        "204":
          description: Role updated successfully
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "404":
          description: Role not found

    delete:
      tags:
        - Roles
      summary: Delete a campaign role
      responses:
        "204":
          description: Role deleted successfully
        "401":
          description: Unauthorized
        "404":
          description: Role not found
  /campaigns/{campaign_id}/voters:
    parameters:
      - name: campaign_id
        in: path
        required: true
        schema:
          type: string
          format: number
          pattern: "^[1-9][0-9]*$"

    post:
      tags:
        - Voters
      summary: Add voters to a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVotersRequest"
      responses:
        "201":
          description: Voters added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateVotersResponse"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "404":
          description: Campaign not found
        "409":
          description: One or more voters already exists in campaign

    get:
      tags:
        - Voters
      summary: Get all voters in a campaign
      responses:
        "200":
          description: Voters retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voter"
        "401":
          description: Unauthorized
        "404":
          description: Campaign not found
  /campaigns/{campaign_id}/voters/delete:
    parameters:
      - name: campaign_id
        in: path
        required: true
        schema:
          type: string
          format: number
          pattern: "^[1-9][0-9]*$"

    post:
      tags:
        - Voters
      summary: Delete 1 or more voters from a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteVotersRequest"
      responses:
        "204":
          description: Voters deleted successfully
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "404":
          description: Campaign not found
  /campaigns/{campaign_id}/voters/invite:
    parameters:
      - name: campaign_id
        in: path
        required: true
        schema:
          type: string
          format: number
          pattern: "^[1-9][0-9]*$"

    post:
      tags:
        - Voters
      summary: Send voting invitations
      description: >
        Sends voting invitations to all voters in the campaign who have not yet been sent an invitation.
      responses:
        "202":
          description: Invitations accepted for sending
        "401":
          description: Unauthorized
        "404":
          description: Campaign not found
  /vote/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Voting
      summary: Get campaign information for a voter
      description: >
        Returns the campaign, roles, and candidates associated with
        the voter's voting token.
      responses:
        "200":
          description: Voting information retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VotingCampaign"
        "404":
          description: Voting token not found or invalid
        "409":
          description: Voting is not currently open
...
components:
  schemas:
...
    CreateRoleRequest:
      type: object
      required:
        - title
        - description
        - no_of_positions
        - enable_abstention
      properties:
        title:
          type: string
        description:
          type: string
        no_of_positions:
          type: integer
          minimum: 1
        enable_abstention:
          type: boolean
    CreateRoleResponse:
      type: object
      required:
        - role_id
      properties:
        role_id:
          type: string
          format: number
          pattern: "^[1-9][0-9]*$"
    UpdateRoleRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
        description:
          type: string
        no_of_positions:
          type: integer
          minimum: 1
        enable_abstention:
          type: boolean
    Role:
      type: object
      required:
        - role_id
        - title
        - description
        - no_of_positions
        - enable_abstention
      properties:
        role_id:
          type: string
          format: number
          pattern: "^[1-9][0-9]*$"
        title:
          type: string
        description:
          type: string
        no_of_positions:
          type: integer
          minimum: 1
        enable_abstention:
          type: boolean
    CreateVotersRequest:
      type: object
      required:
        - emails
      properties:
        emails:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            format: email
    DeleteVotersRequest:
      type: object
      required:
        - voter_ids
      properties:
        voter_ids:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            format: number
            pattern: "^[1-9][0-9]*$"
    Voter:
      type: object
      required:
        - voter_id
        - email
        - status
      properties:
        voter_id:
          type: string
          format: number
          pattern: "^[1-9][0-9]*$"
        email:
          type: string
          format: email
        status:
          type: string
          enum: [pending, invited, voted]
    VotingCampaign:
      type: object
      required:
        - campaign_id
        - title
        - description
        - opening_date_time
        - closing_date_time
        - roles
      properties:
        campaign_id:
          type: string
          format: number
          pattern: "^[1-9][0-9]*$"
        title:
          type: string
        description:
          type: string
        opening_date_time:
          type: string
          format: date-time
        closing_date_time:
          type: string
          format: date-time
        roles:
          type: array
          items:
            $ref: "#/components/schemas/VotingRole"
    CreateVotersResponse:
      type: object
      required:
        - voter_ids
      properties:
        voter_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: number
            pattern: "^[1-9][0-9]*$"
    VotingRole:
      type: object
      required:
        - role_id
        - title
        - description
        - no_of_positions
        - enable_abstention
        - candidates
      properties:
        role_id:
          type: string
          format: number
          pattern: "^[1-9][0-9]*$"
        title:
          type: string
        description:
          type: string
        no_of_positions:
          type: integer
          minimum: 1
        enable_abstention:
          type: boolean
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/VotingCandidate"
    VotingCandidate:
      type: object
      required:
        - candidate_id
        - first_name
        - last_name
      properties:
        candidate_id:
          type: string
          format: number
          pattern: "^[1-9][0-9]*$"
        first_name:
          type: string
        last_name:
          type: string
        manifesto:
          type: string
...